+++ buession-redis/src/main/java/com/buession/redis/core/Command.java
@@ -29,23 +29,76 @@ import com.buession.redis.core.command.ProtocolCommand;
 import com.buession.redis.exception.RedisException;
 
 /**
+ * Redis 命令接口
+ *
+ * @param <R>
+ * 		命令执行结果类型
+ *
  * @author Yong.Teng
  * @since 2.0.0
  */
 public interface Command<R> {
 
+	/**
+	 * 返回协议命令
+	 *
+	 * @return 协议命令
+	 */
 	ProtocolCommand getCommand();
 
+	/**
+	 * 执行 Redis 命令
+	 *
+	 * @return Redis 命令执行结果
+	 *
+	 * @throws RedisException
+	 * 		Redis 命令执行异常
+	 */
 	R execute() throws RedisException;
 
+	/**
+	 * 运行 Redis 命令
+	 *
+	 * @return Redis 命令运行结果
+	 *
+	 * @throws RedisException
+	 * 		Redis 命令运行异常
+	 */
 	default R run() throws RedisException{
 		return run(null);
 	}
 
+	/**
+	 * 运行 Redis 命令
+	 *
+	 * @param arguments
+	 * 		Redis 命令参数
+	 *
+	 * @return Redis 命令运行结果
+	 *
+	 * @throws RedisException
+	 * 		Redis 命令运行异常
+	 */
 	R run(final CommandArguments arguments) throws RedisException;
 
+	/**
+	 * Redis 命令运行器
+	 *
+	 * @author Yong.Teng
+	 */
 	interface Runner {
 
+		/**
+		 * 运行 Redis 命令
+		 *
+		 * @param <R>
+		 * 		命令运行结果类型
+		 *
+		 * @return Redis 命令运行结果
+		 *
+		 * @throws RedisException
+		 * 		Redis 命令运行异常
+		 */
 		<R> R run() throws RedisException;
 
 	}